breaking things to reusable components
