+++ pages/index.js
@@ -1,21 +1,13 @@
 import { buildQuery, arweave, createPostInfo, getRandomEmoji } from '../utils'
 import { useState, useEffect, useContext, useRef } from 'react'
 import { css } from '@emotion/css'
-import formatDistance from 'date-fns/formatDistance'
 import { BundlrContext } from '../context'
-import ReactMarkdown from 'react-markdown'
 import Link from 'next/link'
 
-const topics = {
-  gm: 'gm',
-  degen: 'degen',
-  nfts: 'nfts',
-  defi: 'defi',
-  daos: 'daos',
-  web3: 'web3',
-  solidity: 'solidity',
-  events: 'events'
-}
+import Topics from '../components/Topics.component'
+import Button from '../components/Button.component'
+import PostListContainer from '../components/PostListContainer.component'
+import PostInputComponent from '../components/PostInput.component'
 
 function App() {
   const [postInfos, setPostInfos] = useState([])
@@ -63,7 +55,7 @@ function App() {
       topic: topicToSave ? topicToSave : null,
       emoji: getRandomEmoji()
     }
-  
+
     let tx = await bundlrInstance.createTransaction(JSON.stringify(post), { tags })
 
     try {
@@ -100,7 +92,7 @@ function App() {
     const results = await arweave.api.post('/graphql', query)
       .catch(err => {
         console.error('GraphQL query failed')
-         throw new Error(err);
+        throw new Error(err);
       });
     const edges = results.data.data.transactions.edges
     const posts = await Promise.all(
@@ -112,52 +104,14 @@ function App() {
     console.log('posts: ', sorted)
   }
 
-  function onChange(e) {
-    setPostInput(e.target.value)
-  }
-
-  function checkTopicFilterStyle(topic) {
-    if (topic === topicToFilter) {
-      return css`
-        background-color: #ff5d9d !important;
-      `
-    }
-  }
-
-  function setFilter(topic) {
-    if (topicToFilter === topic) {
-      setTopicToFilter('')
-      getPostInfo()
-    } else {
-      setTopicToFilter(topic)
-      getPostInfo(topic)
-    }
-  }
-
-  function checkTopicStyle(topic) {
-    if (topic === topicToSave) {
-      return css`
-        background-color: #b84cff !important;
-      `
-    }
-  }
-
-  function setTopic(topic) {
-    if (topicToSave === topic) {
-      setTopicToSave('')
-    } else {
-      setTopicToSave(topic)
-    }
-  }
-
   const balanceZero = balance === "0.0"
-  
+
   return (
     <div>
       {
         !bundlrInstance ? (
           <div className={horizontalPaddingStyle}>
-             <button className={button} onClick={connectCeramic}>Connect wallet to start chatting!</button>
+            <Button handleClick={connectCeramic}>Connect wallet to start chating!</Button>
           </div>
         ) : balanceZero ? (
           <div className={horizontalPaddingStyle}>
@@ -168,26 +122,7 @@ function App() {
               </Link>
             </h4>
           </div>
-        ) : (
-          <div className={postInputContainerStyle}>
-            <input
-              onChange={onChange}
-              placeholder="New post"
-              value={postInput}
-              className={postInputStyle}
-            />
-            <div className={filtersListStyle}>
-            {
-              Object.keys(topics).map((topic, i) => (
-                <div key={i} onClick={() => setTopic(topic)}>
-                  <p className={checkTopicStyle(topic)}>{topic}</p>
-                </div>
-              ))
-            }
-            </div>
-            <button className={button} onClick={createPost}>Send</button>
-          </div>
-        )
+        ) : <PostInputComponent handleClick={createPost} />
       }
       {
         isLoading && <h1 className={filtersContainerStyle}>Loading chat...</h1>
@@ -195,126 +130,17 @@ function App() {
       {
         !isLoading && (
           <>
-            <div className={filtersContainerStyle}>
-              <div className={filterheaderStyle}>
-                <h3>Filter by topic</h3>
-                <img onClick={() => getPostInfo()} className={refreshImageStyle} src="/refresh.svg" />
-              </div>
-              <div className={filtersListStyle}>
-                {
-                  Object.keys(topics).map((topic, i) => (
-                    <div key={i}>
-                      <p className={checkTopicFilterStyle(topic)} onClick={() => setFilter(topic)}>{topic}</p>
-                    </div>
-                  ))
-                }
-              </div>
-            </div>
-            <div className={postListContainer}>
-            {
-              postInfos.map((post, i) => (
-                post.request.data.post && (
-                  <div className={postWrapper} key={i} >
-                    {
-                      post.request.data.profileImage ? (
-                        <img
-                          src={post.request.data.profileImage}
-                          className={profileImageStyle}
-                        />
-                      ) : (
-                        <p className={emojiStyle}>{post.request.data.emoji ? post.request.data.emoji : getRandomEmoji()}</p>
-                      )
-                    }
-                    <div className={postContainer}>
-                      <p>{post.request.data.username ? post.request.data.username : post.request.data.createdBy}</p>
-                      <ReactMarkdown>{post.request.data.post}</ReactMarkdown>
-                      <div className={timeAndTopicContainerStyle}>
-                        <p>{formatDistance(new Date(), new Date(post.request.data.createdAt)) + ' ago'}</p>
-                        {
-                          post.request.data.topic && <p>{post.request.data.topic}</p>
-                        }
-                      </div>
-                    </div>
-                  </div>
-                )
-              ))
-            }
-            </div>
+            <Topics getPostInfo={getPostInfo} />
+            <PostListContainer postInfos={postInfos} />
           </>
-          )
-        }
+        )
+      }
     </div>
   );
 }
 
 const horizontalPaddingStyle = css`
   padding: 4px 30px;
-`
-
-const postWrapper = css`
-  border-bottom: 1px solid rgba(0, 0, 0, .2);
-  display: flex;
-  align-items: flex-start;
-  border-top: 1px solid rgba(255, 255, 255, .1);
-  ${horizontalPaddingStyle},
-  h1 {
-    font-size: 40px;
-    margin: 20px 30px 0px 0px;
-  }
-`
-
-const postContainer = css`
-  margin-bottom: 10px;
-  p:first-child {
-    color: #9e54b9;
-    font-size: 22px;
-    margin-bottom: 0px;
-
-  }
-  p:nth-child(2) {
-    font-size: 29px;
-    margin: 17px 0px;
-    font-size: 28px;
-  }
-`
-
-const postInputContainerStyle = css`
-  ${horizontalPaddingStyle},
-  display: flex;
-  flex-direction: column;
-  width: 300px;
-`
-
-const postInputStyle = css`
-  padding: 12px;
-  font-size: 22px;
-  min-width: 250px;
-  border-radius: 7px;
-  border-color: rgba(0, 0, 0, .05);
-  &:focus {
-    border-color: rgba(0, 0, 0, .1);
-    outline: none;
-  }
-`
-
-const button = css`
-  color: white;
-  background-image: linear-gradient(120deg, #ff266a 0%, #c926ff 50%, #3d04cd 100%);
-  padding: 13px 35px;
-  border-radius: 7px;
-  border: none;
-  outline: none;
-  box-shadow: 0 6px 20px rgba(255, 38, 106, .15);
-  cursor: pointer;
-  transition: all .3s;
-  &:hover {
-    box-shadow: 0 6px 20px rgba(255, 38, 106, .3);
-  }
-`
-
-const postListContainer = css`
-  margin-top: 10px;
-  border-top: 1px solid black;
 `
 
 const filtersContainerStyle = css`
@@ -325,66 +151,12 @@ const filtersContainerStyle = css`
   }
 `
 
-const filterheaderStyle = css`
-  display: flex;
-  align-items: flex-start;
-  margin-top: 20px;
-  h3 {
-    margin: 0px;
-  }
-  img {
-    margin-left: 15px;
-    cursor: pointer;
-  }
-`
-
-const profileImageStyle = css`
-  width: 56px;
-  height: 56px;
-  object-fit: cover;
-  border-radius: 28px;
-  margin-top: 26px;
-  margin-right: 12px;
-`
-
-const emojiStyle = css`
-  font-size: 40px;
-  margin-right: 26px;
-`
-
-const filtersListStyle = css`
-  display: flex;
-  p {
-    background-color: #1d1a27;
-    border: 1px solid #302c3f;
-    color: white;
-    margin-right: 6px;
-    padding: 8px 24px;
-    border-radius: 27px;
-    cursor: pointer;
-    font-weight: bold;
-  }
-`
-
 const fundWalletLinkStyle = css`
-  color: #b84cff;
+color: #b84cff;
 `
 
 const refreshImageStyle = css`
-  width: 30px;
-`
-
-const timeAndTopicContainerStyle = css`
-  display: flex;
-  p:first-child {
-    margin: 0;
-  }
-  p:nth-child(2) {
-    margin: 0;
-    margin-left: 10px;
-    font-size: 22px;
-    color: #ff5d9d;
-  }
+width: 30px;
 `
 
 export default App;
